pyproject.toml
[project]
name = "graphrag-developer-challenge3"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "neo4j-graphrag[openai] (>=1.10.1,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-neo4j (>=0.6.0,<0.7.0)",
    "langchain (>=1.0.3)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langchain-openai (>=1.0.1,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "cohere (>=5.20.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.1.0,<8.0.0)",
    "notebook (>=7.4.7,<8.0.0)"
]
